+++ TodoAPIAssignment.DataAccessLibrary.Tests/UnitTests/TodoUnitTests/DeleteTodoUnitTests.cs
@@ -3,7 +3,7 @@ using Microsoft.EntityFrameworkCore;
 using TodoAPIAssignment.DataAccessLibrary.Enums;
 using TodoAPIAssignment.DataAccessLibrary.Models;
 
-namespace TodoAPIAssignment.DataAccessLibrary.Tests.UnitTests;
+namespace TodoAPIAssignment.DataAccessLibrary.Tests.UnitTests.TodoUnitTests;
 
 [TestFixture]
 [Category("Unit")]
@@ -70,8 +70,8 @@ public class DeleteTodoUnitTests
     }
 
     [TearDown]
-    public void TearDown() 
-    { 
-    
+    public void TearDown()
+    {
+
     }
 }
